make log2 more type generic
